+++ src/raters/stars/star.vue
@@ -4,7 +4,9 @@ export default Polygon.extend({
     data() {
         return {
             points: [19.8, 2.2, 6.6, 43.56, 39.6, 17.16, 0, 17.16, 33, 43.56],
-            base: 43
+            originalWidth: 43,
+            originalHeight: 43,
+            borders: 3
         }
     }
 });