+++ src/reg/font-awesome-rating.js
@@ -1,0 +1,3 @@
+import FontAwesomeRating from '../raters/font-awesome/font-awesome-rating.vue';
+
+Vue.component('fa-rating', FontAwesomeRating);